Azure-Search/requirements.txt
# FastAPI and Uvicorn for the backend server
fastapi
uvicorn

# Azure SDKs for various Azure services
azure-identity
azure-search-documents
azure-ai-formrecognizer
azure-storage-blob

# OpenAI SDK for generating answers
openai

# Python utilities for environment management
python-dotenv

# Transformers and PyTorch for embeddings
transformers
torch

# Libraries for document handling
docx

# Streamlit for the UI
streamlit

# Requests for making HTTP requests
requests
